removed PCAModel object and moved its methods to StatisticalMeshModel object
