+++ src/main/scala/scalismo/statisticalmodel/StatisticalMeshModel.scala
@@ -16,12 +16,16 @@
 package scalismo.statisticalmodel
 
 import breeze.linalg.{ DenseMatrix, DenseVector }
-import scalismo.common.{VectorField, PointId, DiscreteVectorField}
+import scalismo.common.{ VectorField, PointId, DiscreteVectorField }
 import scalismo.geometry.{ Point, _3D }
 import scalismo.mesh.{ Mesh, TriangleMesh }
+import scalismo.numerics.FixedPointsUniformMeshSampler3D
 import scalismo.registration.{ RigidTransformation, Transformation }
 import scalismo.statisticalmodel.DiscreteLowRankGaussianProcess.Eigenpair
 import scalismo.mesh.TriangleCell
+import scalismo.statisticalmodel.dataset.DataCollection
+
+import scala.util.{ Success, Failure, Try }
 
 /**
  * A StatisticalMeshModel is isomorphic to a [[DiscreteLowRankGaussianProcess]]. The difference is that while the DiscreteLowRankGaussianProcess
@@ -204,9 +208,36 @@ object StatisticalMeshModel {
    * Creates a new DiscreteLowRankGaussianProcess, where the mean and covariance matrix are estimated from the given transformations.
    *
    */
-  def createWithPCA(referenceMesh: TriangleMesh, fields: Seq[VectorField[_3D,_3D]]): StatisticalMeshModel = {
+  def createWithPCA(referenceMesh: TriangleMesh, fields: Seq[VectorField[_3D, _3D]]): StatisticalMeshModel = {
     val dgp = DiscreteLowRankGaussianProcess.createWithPCA(referenceMesh, fields)
     new StatisticalMeshModel(referenceMesh, dgp)
+  }
+
+  /**
+   *  Adds a bias model to the given statistical shape model
+   */
+  def augmentModel(model: StatisticalMeshModel, biasModel: GaussianProcess[_3D, _3D], numBasisFunctions: Int): StatisticalMeshModel = {
+
+    val modelGP = model.gp.interpolateNearestNeighbor
+    val newMean = modelGP.mean + biasModel.mean
+    val newCov = modelGP.cov + biasModel.cov
+    val newGP = GaussianProcess(newMean, newCov)
+    val sampler = FixedPointsUniformMeshSampler3D(model.referenceMesh, 2 * numBasisFunctions, 42)
+    val newLowRankGP = LowRankGaussianProcess.approximateGP(newGP, sampler, numBasisFunctions)
+    StatisticalMeshModel(model.referenceMesh, newLowRankGP)
+  }
+
+  /**
+   * Returns a PCA model with given reference mesh and a set of items in correspondence.
+   * All points of the reference mesh are considered for computing the PCA
+   */
+  def createFromDataCollection(dc: DataCollection): Try[StatisticalMeshModel] = {
+    if (dc.size < 3) return Failure(new Throwable(s"A data collection with at least 3 transformations is required to build a PCA Model (only ${dc.size} were provided)"))
+
+    val fields = dc.dataItems.map { i =>
+      VectorField[_3D, _3D](i.transformation.domain, p => i.transformation(p) - p)
+    }
+    Success(StatisticalMeshModel.createWithPCA(dc.reference, fields))
   }
 
 }
