+++ src/main/scala/scalismo/statisticalmodel/dataset/Crossvalidation.scala
@@ -54,8 +54,8 @@ object Crossvalidation {
     val folds = dc.createCrossValidationFolds(numFolds)
     val evalResultsForFolds = for (fold <- folds) yield {
       val td = fold.trainingData
-      PCAModel.buildModelFromDataCollection(td).map { pcaModel =>
-        val model = if (biasModelAndRank.isDefined) PCAModel.augmentModel(pcaModel, biasModelAndRank.get._1, biasModelAndRank.get._2) else pcaModel
+      StatisticalMeshModel.createFromDataCollection(td).map { pcaModel =>
+        val model = if (biasModelAndRank.isDefined) StatisticalMeshModel.augmentModel(pcaModel, biasModelAndRank.get._1, biasModelAndRank.get._2) else pcaModel
         val evalResults = for (testingItem <- fold.testingData.dataItems) yield {
           val testMesh = dc.reference.transform(testingItem.transformation)
           evalFun(model, testMesh)
